+++ package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
+    "start": "node index.js"
   },
   "engines": {
     "node": "10.x.x"
@@ -32,8 +31,5 @@
     "lodash.uniqby": "^4.7.0",
     "snoostorm": "^1.0.2",
     "universal-analytics": "^0.4.20"
-  },
-  "devDependencies": {
-    "nodemon": "^1.19.3"
   }
 }
